+++ RELEASES.md
@@ -44,6 +44,7 @@ We don't update the MSRV in a patch release, unless we really, really need to.
   - check the in-browser profiler
 * [ ] check the color test
 * [ ] update `eframe_template` and test
+* [ ] update `egui_plot` and test
 * [ ] update `egui_tiles` and test
 * [ ] test with Rerun
 * [ ] `./scripts/check.sh`
@@ -76,7 +77,6 @@ I usually do this all on the `master` branch, but doing it in a release branch i
 (cd crates/ecolor        && cargo publish --quiet)  &&  echo "✅ ecolor"
 (cd crates/epaint        && cargo publish --quiet)  &&  echo "✅ epaint"
 (cd crates/egui          && cargo publish --quiet)  &&  echo "✅ egui"
-(cd crates/egui_plot     && cargo publish --quiet)  &&  echo "✅ egui_plot"
 (cd crates/egui-winit    && cargo publish --quiet)  &&  echo "✅ egui-winit"
 (cd crates/egui_extras   && cargo publish --quiet)  &&  echo "✅ egui_extras"
 (cd crates/egui-wgpu     && cargo publish --quiet)  &&  echo "✅ egui-wgpu"
@@ -94,4 +94,5 @@ I usually do this all on the `master` branch, but doing it in a release branch i
 
 ## After release
 * [ ] publish new `eframe_template`
+* [ ] publish new `egui_plot`
 * [ ] publish new `egui_tiles`